// app.client/src/api/models/LoginRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type LoginRequest = {
    email?: string | null;
    password?: string | null;
    twoFactorCode?: string | null;
    twoFactorRecoveryCode?: string | null;
    token?:  string | null;
};